+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * `Client::list_core_api_resources` added
   * `kube::DynamicResource` exposed at top level
   * Bug: `PatchParams::default_apply()` now requires a manager and renamed to `PatchParams::apply(manager: &str)` for #300
+  * Bug: `DeleteParams` no longer missing for `Api::delete_collection` - #53
 
 0.39.0 / 2020-07-05
 ===================